rest-api/canary/requirements.txt
Django==1.5.1
bottle==0.11.6
distribute==0.6.34
django-tastypie==0.9.15
mimeparse==0.1.3
python-dateutil==2.1
six==1.3.0
uuid==1.30
wsgiref==0.1.2
